Remove duplicate SSG audit rules from supplemental.sh
While the performance impact of duplicate rules is probably minimal,
a duplicate rule can cause rule loading to abort and leave important
events unaudited.

In particular we were seeing that the duplicate MAC-policy watch on
/etc/selinux was halting rule processing after loading only about 80%
of the configured rules.

We also noticed that while remediating a system the SSG would leave
some rules in a file simply named ".rules" that would be excluded when
augenrules compiled the final rules file.

Additionally the SSG is sometimes generating invalid rules with misspelled
exit conditions: EACCESS instead of EACCES and EPRM instead of EPERM.
This bug should be fixed in scap-security-guide 0.1.34.