If path to ISO contains '.' then script fails
If the path to the original ISO file contains a '.' (dot)(e.g. "/home/buchanan/isobuild.d/CentOS-7-x86_64-DVD-1804.iso"), the 'awk -F .' parsing of the MAJOR and MINOR release numbers inserts components of the path instead of just the intended numbers. This causes the output path of the genisoimage command to contain an invalid path. Added a subshell 'cd' to the directory before the 'ls' to prevent the error.